new ActionColumn that has flag for column manager
