+++ src/core/Table/columns/index.ts
@@ -2,4 +2,4 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-export { COLUMN_MANAGER_ID, ManagerColumn } from './managerColumn';
+export { ACTION_CELL_ID, ActionColumn } from './actionColumn';
