+++ src/core/Table/hooks/index.ts
@@ -7,5 +7,4 @@ export { SELECTION_CELL_ID, useSelectionCell } from './useSelectionCell';
 export { useSubRowFiltering } from './useSubRowFiltering';
 export { useSubRowSelection } from './useSubRowSelection';
 export { useResizeColumns } from './useResizeColumns';
-export { useColumnVisibility } from './useColumnVisibility';
 export { useColumnDragAndDrop } from './useColumnDragAndDrop';
